backend/requirements.txt
psycopg2-binary==2.9.5
gunicorn==20.1.0
django==2.2.7
djangorestframework==3.10.3
djangorestframework_simplejwt==4.3.0
django-cors-headers==3.1.1
coreapi==2.3.3
flake8==3.7.9
yamllint==1.18.0
coverage==4.5.4
pillow==9.3.0
